Debugging single ticklists

+++ gw_spaceheat/actors/api_flow_module.py
@@ -460,7 +460,7 @@ class ApiFlowModule(Actor):
             print("1 tick received")
             final_tick_ns = self.nano_timestamps[-1]
             if self.latest_tick_ns is not None:
-                final_nonzero_hz = int(1e9/(final_tick_ns - self.latest_tick_ns))
+                final_nonzero_hz = 1e9/(final_tick_ns - self.latest_tick_ns)
                 print("computing hz:")
                 print(f"Final tick ns - latest tick ns: {final_tick_ns-self.latest_tick_ns}")
                 print(f"freq: {final_nonzero_hz}")
